+++ NEWS.md
@@ -11,6 +11,7 @@
   (#343).
 * `ETS()` will now ignore the smoothing parameter's range when specific 
   parameter value is given (#317).
+* Modified initial parameter values for `ETS()` when bounds = "admissible".
 
 ## Bug fixes